Move services impl out of etcd folder as it only depends on kv.Store
