+++ client/etcd/options_test.go
@@ -24,7 +24,7 @@ import (
 	"testing"
 	"time"
 
-	etcdsd "github.com/m3db/m3cluster/services/client/etcd"
+	"github.com/m3db/m3cluster/services"
 	"github.com/m3db/m3x/instrument"
 
 	"github.com/stretchr/testify/assert"
@@ -80,7 +80,7 @@ func TestOptions(t *testing.T) {
 	opts := NewOptions()
 	assert.Equal(t, "", opts.Zone())
 	assert.Equal(t, "", opts.Env())
-	assert.Equal(t, etcdsd.Configuration{}, opts.ServiceDiscoveryConfig())
+	assert.Equal(t, services.NewOptions(), opts.ServicesOptions())
 	assert.Equal(t, "", opts.CacheDir())
 	assert.Equal(t, "", opts.Service())
 	assert.Equal(t, []Cluster{}, opts.Clusters())
@@ -92,16 +92,17 @@ func TestOptions(t *testing.T) {
 	c2 := NewCluster().SetZone("z2")
 	iopts := instrument.NewOptions().SetReportInterval(time.Minute)
 
+	sdOpts := services.NewOptions().SetInitTimeout(time.Millisecond)
 	opts = opts.SetEnv("env").
 		SetZone("zone").
-		SetServiceDiscoveryConfig(etcdsd.Configuration{}).
+		SetServicesOptions(sdOpts).
 		SetCacheDir("/dir").
 		SetService("app").
 		SetClusters([]Cluster{c1, c2}).
 		SetInstrumentOptions(iopts)
 	assert.Equal(t, "env", opts.Env())
 	assert.Equal(t, "zone", opts.Zone())
-	assert.Equal(t, etcdsd.Configuration{}, opts.ServiceDiscoveryConfig())
+	assert.Equal(t, sdOpts, opts.ServicesOptions())
 	assert.Equal(t, "/dir", opts.CacheDir())
 	assert.Equal(t, "app", opts.Service())
 	assert.Equal(t, 2, len(opts.Clusters()))
