+++ integration/etcd/etcd.go
@@ -31,7 +31,7 @@ import (
 
 	"github.com/m3db/m3cluster/client"
 	etcdclient "github.com/m3db/m3cluster/client/etcd"
-	etcdsd "github.com/m3db/m3cluster/services/client/etcd"
+	"github.com/m3db/m3cluster/services"
 	xclock "github.com/m3db/m3x/clock"
 	"github.com/m3db/m3x/errors"
 
@@ -136,9 +136,7 @@ func (e *embeddedKV) Endpoints() []string {
 func (e *embeddedKV) ConfigServiceClient() (client.Client, error) {
 	eopts := etcdclient.NewOptions().
 		SetInstrumentOptions(e.opts.InstrumentOptions()).
-		SetServiceDiscoveryConfig(etcdsd.Configuration{
-			InitTimeout: e.opts.InitTimeout(),
-		}).
+		SetServicesOptions(services.NewOptions().SetInitTimeout(e.opts.InitTimeout())).
 		SetClusters([]etcdclient.Cluster{
 			etcdclient.NewCluster().SetZone(e.opts.Zone()).SetEndpoints(e.Endpoints()),
 		}).
