+++ services/services.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2016 Uber Technologies, Inc.
+// Copyright (c) 2018 Uber Technologies, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -23,28 +23,611 @@ package services
 import (
 	"errors"
 	"fmt"
-	"os"
+	"sync"
 	"time"
 
 	"github.com/m3db/m3cluster/generated/proto/metadatapb"
 	"github.com/m3db/m3cluster/generated/proto/placementpb"
+	"github.com/m3db/m3cluster/kv"
 	"github.com/m3db/m3cluster/placement"
+	ps "github.com/m3db/m3cluster/placement/service"
+	"github.com/m3db/m3cluster/placement/storage"
 	"github.com/m3db/m3cluster/shard"
-)
+	"github.com/m3db/m3x/log"
+	xwatch "github.com/m3db/m3x/watch"
 
-const (
-	defaultLeaderTimeout = 10 * time.Second
-	defaultResignTimeout = 10 * time.Second
+	"github.com/uber-go/tally"
 )
 
 var (
+	defaultGaugeInterval = 10 * time.Second
+
+	errNoServiceName             = errors.New("no service specified")
+	errNoServiceID               = errors.New("no service id specified")
+	errNoInstanceID              = errors.New("no instance id specified")
+	errAdPlacementMissing        = errors.New("advertisement is missing placement instance")
 	errInstanceNotFound          = errors.New("instance not found")
 	errNilPlacementProto         = errors.New("nil placement proto")
 	errNilPlacementInstanceProto = errors.New("nil placement instance proto")
 	errNilMetadataProto          = errors.New("nil metadata proto")
 )
 
-// NewService creates a new Service
+// NewServices returns a client of Services.
+func NewServices(opts Options) (Services, error) {
+	if err := opts.Validate(); err != nil {
+		return nil, err
+	}
+
+	return &client{
+		opts:           opts,
+		placementKeyFn: keyFnWithNamespace(placementNamespace(opts.NamespaceOptions().PlacementNamespace())),
+		metadataKeyFn:  keyFnWithNamespace(metadataNamespace(opts.NamespaceOptions().MetadataNamespace())),
+		kvManagers:     make(map[string]*kvManager),
+		hbStores:       make(map[string]HeartbeatService),
+		adDoneChs:      make(map[string]chan struct{}),
+		ldSvcs:         make(map[leaderKey]LeaderService),
+		logger:         opts.InstrumentsOptions().Logger(),
+		m:              opts.InstrumentsOptions().MetricsScope(),
+	}, nil
+}
+
+type client struct {
+	sync.RWMutex
+
+	opts           Options
+	placementKeyFn keyFn
+	metadataKeyFn  keyFn
+	kvManagers     map[string]*kvManager
+	hbStores       map[string]HeartbeatService
+	ldSvcs         map[leaderKey]LeaderService
+	adDoneChs      map[string]chan struct{}
+	logger         log.Logger
+	m              tally.Scope
+}
+
+func (c *client) Metadata(sid ServiceID) (Metadata, error) {
+	if err := validateServiceID(sid); err != nil {
+		return nil, err
+	}
+
+	m, err := c.getKVManager(sid.Zone())
+	if err != nil {
+		return nil, err
+	}
+
+	v, err := m.kv.Get(c.metadataKeyFn(sid))
+	if err != nil {
+		return nil, err
+	}
+
+	var mp metadatapb.Metadata
+	if err = v.Unmarshal(&mp); err != nil {
+		return nil, err
+	}
+
+	return NewMetadataFromProto(&mp)
+}
+
+func (c *client) SetMetadata(sid ServiceID, meta Metadata) error {
+	if err := validateServiceID(sid); err != nil {
+		return err
+	}
+
+	m, err := c.getKVManager(sid.Zone())
+	if err != nil {
+		return err
+	}
+
+	mp, err := meta.Proto()
+	if err != nil {
+		return err
+	}
+
+	_, err = m.kv.Set(c.metadataKeyFn(sid), mp)
+	return err
+}
+
+func (c *client) PlacementService(sid ServiceID, opts placement.Options) (placement.Service, error) {
+	if err := validateServiceID(sid); err != nil {
+		return nil, err
+	}
+
+	store, err := c.opts.KVGen()(sid.Zone())
+	if err != nil {
+		return nil, err
+	}
+
+	return ps.NewPlacementService(
+		storage.NewPlacementStorage(store, c.placementKeyFn(sid), opts),
+		opts,
+	), nil
+}
+
+func (c *client) Advertise(ad Advertisement) error {
+	pi := ad.PlacementInstance()
+	if pi == nil {
+		return errAdPlacementMissing
+	}
+
+	if err := validateAdvertisement(ad.ServiceID(), pi.ID()); err != nil {
+		return err
+	}
+
+	m, err := c.Metadata(ad.ServiceID())
+	if err != nil {
+		return err
+	}
+
+	hb, err := c.getHeartbeatService(ad.ServiceID())
+	if err != nil {
+		return err
+	}
+
+	key := adKey(ad.ServiceID(), pi.ID())
+	c.Lock()
+	ch, ok := c.adDoneChs[key]
+	if ok {
+		c.Unlock()
+		return fmt.Errorf("service %s, instance %s is already being advertised", ad.ServiceID(), pi.ID())
+	}
+	ch = make(chan struct{})
+	c.adDoneChs[key] = ch
+	c.Unlock()
+
+	go func() {
+		sid := ad.ServiceID()
+		errCounter := c.serviceTaggedScope(sid).Counter("heartbeat.error")
+
+		tickFn := func() {
+			if isHealthy(ad) {
+				if err := hb.Heartbeat(pi, m.LivenessInterval()); err != nil {
+					c.logger.Errorf("could not heartbeat service %s, %v", sid.String(), err)
+					errCounter.Inc(1)
+				}
+			}
+		}
+
+		tickFn()
+
+		ticker := time.NewTicker(m.HeartbeatInterval())
+		defer ticker.Stop()
+		for {
+			select {
+			case <-ticker.C:
+				tickFn()
+			case <-ch:
+				return
+			}
+		}
+	}()
+	return nil
+}
+
+func (c *client) Unadvertise(sid ServiceID, id string) error {
+	if err := validateAdvertisement(sid, id); err != nil {
+		return err
+	}
+
+	key := adKey(sid, id)
+
+	c.Lock()
+	if ch, ok := c.adDoneChs[key]; ok {
+		// If this client is advertising the instance, stop it.
+		close(ch)
+		delete(c.adDoneChs, key)
+	}
+	c.Unlock()
+
+	hbStore, err := c.getHeartbeatService(sid)
+	if err != nil {
+		return err
+	}
+
+	return hbStore.Delete(id)
+}
+
+func (c *client) Query(sid ServiceID, opts QueryOptions) (Service, error) {
+	if err := validateServiceID(sid); err != nil {
+		return nil, err
+	}
+
+	v, err := c.getPlacementValue(sid)
+	if err != nil {
+		return nil, err
+	}
+
+	service, err := getServiceFromValue(v, sid)
+	if err != nil {
+		return nil, err
+	}
+
+	if !opts.IncludeUnhealthy() {
+		hbStore, err := c.getHeartbeatService(sid)
+		if err != nil {
+			return nil, err
+		}
+
+		ids, err := hbStore.Get()
+		if err != nil {
+			return nil, err
+		}
+
+		service = filterInstances(service, ids)
+	}
+
+	return service, nil
+}
+
+func (c *client) Watch(sid ServiceID, opts QueryOptions) (Watch, error) {
+	if err := validateServiceID(sid); err != nil {
+		return nil, err
+	}
+
+	c.logger.Infof(
+		"adding a watch for service: %s env: %s zone: %s includeUnhealthy: %v",
+		sid.Name(),
+		sid.Environment(),
+		sid.Zone(),
+		opts.IncludeUnhealthy(),
+	)
+
+	kvm, err := c.getKVManager(sid.Zone())
+	if err != nil {
+		return nil, err
+	}
+
+	kvm.RLock()
+	watchable, exist := kvm.serviceWatchables[sid.String()]
+	kvm.RUnlock()
+	if exist {
+		_, w, err := watchable.watch()
+		return w, err
+	}
+
+	// Prepare the watch of placement outside of lock.
+	placementWatch, err := kvm.kv.Watch(c.placementKeyFn(sid))
+	if err != nil {
+		return nil, err
+	}
+
+	initValue, err := c.waitForInitValue(kvm.kv, placementWatch, sid, c.opts.InitTimeout())
+	if err != nil {
+		return nil, fmt.Errorf("could not get init value within timeout, err: %v", err)
+	}
+
+	initService, err := getServiceFromValue(initValue, sid)
+	if err != nil {
+		placementWatch.Close()
+		return nil, err
+	}
+
+	kvm.Lock()
+	defer kvm.Unlock()
+	watchable, exist = kvm.serviceWatchables[sid.String()]
+	if exist {
+		// If a watchable already exist now, we need to clean up the placement watch we just created.
+		placementWatch.Close()
+		_, w, err := watchable.watch()
+		return w, err
+	}
+
+	watchable = newServiceWatchable()
+	sdm := newServiceDiscoveryMetrics(c.serviceTaggedScope(sid))
+
+	if !opts.IncludeUnhealthy() {
+		hbStore, err := c.getHeartbeatService(sid)
+		if err != nil {
+			placementWatch.Close()
+			return nil, err
+		}
+		heartbeatWatch, err := hbStore.Watch()
+		if err != nil {
+			placementWatch.Close()
+			return nil, err
+		}
+		watchable.update(filterInstancesWithWatch(initService, heartbeatWatch))
+		go c.watchPlacementAndHeartbeat(watchable, placementWatch, heartbeatWatch, initValue, sid, initService, sdm.serviceUnmalshalErr)
+	} else {
+		watchable.update(initService)
+		go c.watchPlacement(watchable, placementWatch, initValue, sid, sdm.serviceUnmalshalErr)
+	}
+
+	kvm.serviceWatchables[sid.String()] = watchable
+
+	go updateVersionGauge(placementWatch, sdm.versionGauge)
+
+	_, w, err := watchable.watch()
+	return w, err
+}
+
+func (c *client) HeartbeatService(sid ServiceID) (HeartbeatService, error) {
+	if err := validateServiceID(sid); err != nil {
+		return nil, err
+	}
+
+	return c.getHeartbeatService(sid)
+}
+
+func (c *client) getPlacementValue(sid ServiceID) (kv.Value, error) {
+	kvm, err := c.getKVManager(sid.Zone())
+	if err != nil {
+		return nil, err
+	}
+
+	v, err := kvm.kv.Get(c.placementKeyFn(sid))
+	if err != nil {
+		return nil, err
+	}
+
+	return v, nil
+}
+
+func (c *client) getHeartbeatService(sid ServiceID) (HeartbeatService, error) {
+	c.Lock()
+	defer c.Unlock()
+	hb, ok := c.hbStores[sid.String()]
+	if ok {
+		return hb, nil
+	}
+
+	hb, err := c.opts.HeartbeatGen()(sid)
+	if err != nil {
+		return nil, err
+	}
+
+	c.hbStores[sid.String()] = hb
+	return hb, nil
+}
+
+func (c *client) LeaderService(sid ServiceID, opts ElectionOptions) (LeaderService, error) {
+	if sid == nil {
+		return nil, errNoServiceID
+	}
+
+	if opts == nil {
+		opts = NewElectionOptions()
+	}
+
+	key := leaderCacheKey(sid, opts)
+
+	c.RLock()
+	if ld, ok := c.ldSvcs[key]; ok {
+		c.RUnlock()
+		return ld, nil
+	}
+	c.RUnlock()
+
+	c.Lock()
+	defer c.Unlock()
+
+	if ld, ok := c.ldSvcs[key]; ok {
+		return ld, nil
+	}
+
+	ld, err := c.opts.LeaderGen()(sid, opts)
+	if err != nil {
+		return nil, err
+	}
+
+	c.ldSvcs[key] = ld
+	return ld, nil
+}
+
+func (c *client) getKVManager(zone string) (*kvManager, error) {
+	c.Lock()
+	defer c.Unlock()
+	m, ok := c.kvManagers[zone]
+	if ok {
+		return m, nil
+	}
+
+	kv, err := c.opts.KVGen()(zone)
+	if err != nil {
+		return nil, err
+	}
+
+	m = &kvManager{
+		kv:                kv,
+		serviceWatchables: map[string]serviceWatchable{},
+	}
+
+	c.kvManagers[zone] = m
+	return m, nil
+}
+
+func (c *client) watchPlacement(
+	w serviceWatchable,
+	vw kv.ValueWatch,
+	initValue kv.Value,
+	sid ServiceID,
+	errCounter tally.Counter,
+) {
+	for range vw.C() {
+		newService := c.serviceFromUpdate(vw.Get(), initValue, sid, errCounter)
+		if newService == nil {
+			continue
+		}
+
+		w.update(newService)
+	}
+}
+
+func (c *client) watchPlacementAndHeartbeat(
+	w serviceWatchable,
+	vw kv.ValueWatch,
+	heartbeatWatch xwatch.Watch,
+	initValue kv.Value,
+	sid ServiceID,
+	service Service,
+	errCounter tally.Counter,
+) {
+	for {
+		select {
+		case <-vw.C():
+			newService := c.serviceFromUpdate(vw.Get(), initValue, sid, errCounter)
+			if newService == nil {
+				continue
+			}
+
+			service = newService
+		case <-heartbeatWatch.C():
+			c.logger.Infof("received heartbeat update")
+		}
+		w.update(filterInstancesWithWatch(service, heartbeatWatch))
+	}
+}
+
+func (c *client) serviceFromUpdate(
+	value kv.Value,
+	initValue kv.Value,
+	sid ServiceID,
+	errCounter tally.Counter,
+) Service {
+	if value == nil {
+		// NB(cw) this can only happen when the placement has been deleted
+		// it is safer to let the user keep using the old topology
+		c.logger.Info("received placement update with nil value")
+		return nil
+	}
+
+	if initValue != nil && !value.IsNewer(initValue) {
+		// NB(cw) this can only happen when the init wait called a Get() itself
+		// so the init value did not come from the watch, when the watch gets created
+		// the first update from it may from the same version.
+		c.logger.Infof("received stale placement update on version %d, skip", value.Version())
+		return nil
+	}
+
+	newService, err := getServiceFromValue(value, sid)
+	if err != nil {
+		c.logger.Errorf("could not unmarshal update from kv store for placement on version %d, %v", value.Version(), err)
+		errCounter.Inc(1)
+		return nil
+	}
+
+	c.logger.Infof("successfully parsed placement on version %d", value.Version())
+	return newService
+}
+
+func (c *client) serviceTaggedScope(sid ServiceID) tally.Scope {
+	return c.m.Tagged(
+		map[string]string{
+			"sd_service": sid.Name(),
+			"sd_env":     sid.Environment(),
+			"sd_zone":    sid.Zone(),
+		},
+	)
+}
+
+func isHealthy(ad Advertisement) bool {
+	healthFn := ad.Health()
+	return healthFn == nil || healthFn() == nil
+}
+
+func filterInstances(s Service, ids []string) Service {
+	if s == nil {
+		return nil
+	}
+
+	instances := make([]ServiceInstance, 0, len(s.Instances()))
+	for _, id := range ids {
+		if instance, err := s.Instance(id); err == nil {
+			instances = append(instances, instance)
+		}
+	}
+
+	return NewService().
+		SetInstances(instances).
+		SetSharding(s.Sharding()).
+		SetReplication(s.Replication())
+}
+
+func filterInstancesWithWatch(s Service, hbw xwatch.Watch) Service {
+	if hbw.Get() == nil {
+		return s
+	}
+	return filterInstances(s, hbw.Get().([]string))
+}
+
+func updateVersionGauge(vw kv.ValueWatch, versionGauge tally.Gauge) {
+	for range time.Tick(defaultGaugeInterval) {
+		v := vw.Get()
+		if v != nil {
+			versionGauge.Update(float64(v.Version()))
+		}
+	}
+}
+
+func getServiceFromValue(value kv.Value, sid ServiceID) (Service, error) {
+	p, err := placementFromValue(value)
+	if err != nil {
+		return nil, err
+	}
+
+	return NewServiceFromPlacement(p, sid), nil
+}
+
+func (c *client) waitForInitValue(kvStore kv.Store, w kv.ValueWatch, sid ServiceID, timeout time.Duration) (kv.Value, error) {
+	if timeout <= 0 {
+		timeout = defaultInitTimeout
+	}
+	select {
+	case <-w.C():
+		return w.Get(), nil
+	case <-time.After(timeout):
+		return kvStore.Get(c.placementKeyFn(sid))
+	}
+}
+
+func validateAdvertisement(sid ServiceID, id string) error {
+	if sid == nil {
+		return errNoServiceID
+	}
+
+	if id == "" {
+		return errNoInstanceID
+	}
+
+	return nil
+}
+
+// cache key for leader service clients
+type leaderKey struct {
+	sid           string
+	leaderTimeout time.Duration
+	resignTimeout time.Duration
+	ttl           int
+}
+
+func leaderCacheKey(sid ServiceID, opts ElectionOptions) leaderKey {
+	return leaderKey{
+		sid:           sid.String(),
+		leaderTimeout: opts.LeaderTimeout(),
+		resignTimeout: opts.ResignTimeout(),
+		ttl:           opts.TTLSecs(),
+	}
+}
+
+type kvManager struct {
+	sync.RWMutex
+
+	kv                kv.Store
+	serviceWatchables map[string]serviceWatchable
+}
+
+func newServiceDiscoveryMetrics(m tally.Scope) serviceDiscoveryMetrics {
+	return serviceDiscoveryMetrics{
+		versionGauge:        m.Gauge("placement.version"),
+		serviceUnmalshalErr: m.Counter("placement.unmarshal.error"),
+	}
+}
+
+type serviceDiscoveryMetrics struct {
+	versionGauge        tally.Gauge
+	serviceUnmalshalErr tally.Counter
+}
+
+// NewService creates a new Service.
 func NewService() Service { return new(service) }
 
 // NewServiceFromProto takes the data from a placement and a service id and
@@ -109,7 +692,7 @@ func (s *service) SetInstances(insts []ServiceInstance) Service { s.instances = 
 func (s *service) SetReplication(r ServiceReplication) Service  { s.replication = r; return s }
 func (s *service) SetSharding(ss ServiceSharding) Service       { s.sharding = ss; return s }
 
-// NewServiceReplication creates a new ServiceReplication
+// NewServiceReplication creates a new ServiceReplication.
 func NewServiceReplication() ServiceReplication { return new(serviceReplication) }
 
 type serviceReplication struct {
@@ -119,7 +702,7 @@ type serviceReplication struct {
 func (r *serviceReplication) Replicas() int                          { return r.replicas }
 func (r *serviceReplication) SetReplicas(rep int) ServiceReplication { r.replicas = rep; return r }
 
-// NewServiceSharding creates a new ServiceSharding
+// NewServiceSharding creates a new ServiceSharding.
 func NewServiceSharding() ServiceSharding { return new(serviceSharding) }
 
 type serviceSharding struct {
@@ -132,7 +715,7 @@ func (s *serviceSharding) IsSharded() bool                     { return s.isShar
 func (s *serviceSharding) SetNumShards(n int) ServiceSharding  { s.numShards = n; return s }
 func (s *serviceSharding) SetIsSharded(v bool) ServiceSharding { s.isSharded = v; return s }
 
-// NewServiceInstance creates a new ServiceInstance
+// NewServiceInstance creates a new ServiceInstance.
 func NewServiceInstance() ServiceInstance { return new(serviceInstance) }
 
 // NewServiceInstanceFromProto creates a new service instance from proto.
@@ -186,7 +769,7 @@ func (i *serviceInstance) SetServiceID(service ServiceID) ServiceInstance {
 	return i
 }
 
-// NewAdvertisement creates a new Advertisement
+// NewAdvertisement creates a new Advertisement.
 func NewAdvertisement() Advertisement { return new(advertisement) }
 
 type advertisement struct {
@@ -205,7 +788,7 @@ func (a *advertisement) SetPlacementInstance(p placement.Instance) Advertisement
 	return a
 }
 
-// NewServiceID creates new ServiceID
+// NewServiceID creates new ServiceID.
 func NewServiceID() ServiceID { return new(serviceID) }
 
 type serviceID struct {
@@ -224,17 +807,7 @@ func (sid *serviceID) String() string {
 	return fmt.Sprintf("[name: %s, env: %s, zone: %s]", sid.name, sid.env, sid.zone)
 }
 
-// NewQueryOptions creates new QueryOptions
-func NewQueryOptions() QueryOptions { return new(queryOptions) }
-
-type queryOptions struct {
-	includeUnhealthy bool
-}
-
-func (qo *queryOptions) IncludeUnhealthy() bool                  { return qo.includeUnhealthy }
-func (qo *queryOptions) SetIncludeUnhealthy(h bool) QueryOptions { qo.includeUnhealthy = h; return qo }
-
-// NewMetadata creates new Metadata
+// NewMetadata creates new Metadata.
 func NewMetadata() Metadata { return new(metadata) }
 
 // NewMetadataFromProto converts a Metadata proto message to an instance of
@@ -284,118 +857,4 @@ func (m *metadata) Proto() (*metadatapb.Metadata, error) {
 		LivenessInterval:  int64(m.LivenessInterval()),
 		HeartbeatInterval: int64(m.HeartbeatInterval()),
 	}, nil
-}
-
-// NewElectionOptions returns an empty ElectionOptions.
-func NewElectionOptions() ElectionOptions {
-	eo := electionOpts{
-		leaderTimeout: defaultLeaderTimeout,
-		resignTimeout: defaultResignTimeout,
-	}
-
-	return eo
-}
-
-type electionOpts struct {
-	leaderTimeout time.Duration
-	resignTimeout time.Duration
-	ttlSecs       int
-}
-
-func (e electionOpts) LeaderTimeout() time.Duration {
-	return e.leaderTimeout
-}
-
-func (e electionOpts) SetLeaderTimeout(t time.Duration) ElectionOptions {
-	e.leaderTimeout = t
-	return e
-}
-
-func (e electionOpts) ResignTimeout() time.Duration {
-	return e.resignTimeout
-}
-
-func (e electionOpts) SetResignTimeout(t time.Duration) ElectionOptions {
-	e.resignTimeout = t
-	return e
-}
-
-func (e electionOpts) TTLSecs() int {
-	return e.ttlSecs
-}
-
-func (e electionOpts) SetTTLSecs(ttl int) ElectionOptions {
-	e.ttlSecs = ttl
-	return e
-}
-
-type campaignOpts struct {
-	val string
-}
-
-// NewCampaignOptions returns an empty CampaignOptions.
-func NewCampaignOptions() (CampaignOptions, error) {
-	h, err := os.Hostname()
-	if err != nil {
-		return nil, err
-	}
-
-	return campaignOpts{val: h}, nil
-}
-
-func (c campaignOpts) LeaderValue() string {
-	return c.val
-}
-
-func (c campaignOpts) SetLeaderValue(v string) CampaignOptions {
-	c.val = v
-	return c
-}
-
-type options struct {
-	namespaceOpts NamespaceOptions
-}
-
-// NewOptions constructs a new Options.
-func NewOptions() Options {
-	return &options{
-		namespaceOpts: NewNamespaceOptions(),
-	}
-}
-
-func (o options) NamespaceOptions() NamespaceOptions {
-	return o.namespaceOpts
-}
-
-func (o options) SetNamespaceOptions(opts NamespaceOptions) Options {
-	o.namespaceOpts = opts
-	return o
-}
-
-type namespaceOpts struct {
-	placement string
-	metadata  string
-}
-
-// NewNamespaceOptions constructs a new NamespaceOptions.
-func NewNamespaceOptions() NamespaceOptions {
-	return &namespaceOpts{}
-}
-
-func (opts namespaceOpts) PlacementNamespace() string {
-	return opts.placement
-}
-
-func (opts namespaceOpts) SetPlacementNamespace(v string) NamespaceOptions {
-	opts.placement = v
-	return opts
-}
-
-func (opts namespaceOpts) MetadataNamespace() string {
-	return opts.metadata
-}
-
-func (opts namespaceOpts) SetMetadataNamespace(v string) NamespaceOptions {
-	opts.metadata = v
-	return opts
 }
